add access control and file uploading mecahnism

+++ app/Http/Controllers/PostsController.php
@@ -3,11 +3,17 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 use DB;
 
 class PostsController extends Controller
 {
+    public function __construct()
+    {
+        // prevent guest from accessing the post, except certain views
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
 
     /**
      * Display a listing of the resource.
@@ -52,13 +58,31 @@ class PostsController extends Controller
     {
         $this->validate($request,[
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999',
         ]);
         
+        // handle file upload
+        if($request->hasFile('cover_image')) {
+            // get filename with the extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // get just file name
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // get just extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // filename to store
+            $filenameToStore = $filename.'_'.time().'.'.$extension;
+            //upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_image', $filenameToStore);
+        } else {
+            $filenameToStore = 'noimage.jpg';
+        }
+
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->user_id = auth()->user()->id;
+        $post->cover_image = $filenameToStore;
         $post->save();
         return redirect('/posts')->with('success', 'Post Created'); 
     }
@@ -84,6 +108,12 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post =  Post::find($id);
+
+        // check for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+
         return view('posts/edit')->with('post', $post);
     }
 
@@ -100,10 +130,26 @@ class PostsController extends Controller
             'title' => 'required',
             'body' => 'required'
         ]);
+
+        if($request->hasFile('cover_image')) {
+            // get filename with the extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // get just file name
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // get just extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // filename to store
+            $filenameToStore = $filename.'_'.time().'.'.$extension;
+            //upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_image', $filenameToStore);
+        }
         
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        if ($request->hasFile('cover_image')) {
+            $post->cover_image = $filenameToStore;
+        } 
         $post->save();
         return redirect('/posts')->with('success', 'Post Updated');
     }
@@ -117,6 +163,17 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        // check for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+
+        if ($post->cover_image != 'noimage.jpg') {
+            // delete image
+            Storage::delete('/LaravelProjectDemo/public/cover_image/'.$post->cover_image);
+        }
+
         $post->delete();
         return redirect('/posts')->with('success', 'Post removed');
     }
